route: executive summary as landing page
